requirements.txt
google-generativeai
tqdm
Pillow
beautifulsoup4
lxml
tkextrafont
opencv-python
numpy